+++ app/src/main/res/layout/addrepeatslistitem.xml
@@ -60,7 +60,9 @@
         android:layout_alignParentLeft="true"
         android:layout_toStartOf="@id/addImage"
         android:layout_below="@id/imageView"
-        android:layout_margin="5dp"/>
+        android:layout_margin="5dp"
+        android:singleLine="true"
+        android:imeOptions="actionNext"/>
 
     <EditText
         android:id="@+id/answerBox"
@@ -72,5 +74,7 @@
         android:layout_alignParentStart="true"
         android:layout_alignParentLeft="true"
         android:layout_toStartOf="@id/addImage"
-        android:layout_margin="5dp"/>
+        android:layout_margin="5dp"
+        android:singleLine="true"
+        android:imeOptions="actionNext"/>
 </RelativeLayout>
