+++ app/src/main/res/layout/test_item.xml
@@ -13,7 +13,7 @@
         android:text="TextView"
         android:textSize="20sp"
         android:layout_margin="12dp"
-        android:layout_alignParentTop="true"/>
+        android:layout_alignParentTop="true" />
 
     <ImageView
         android:id="@+id/imageViewTest"
@@ -32,6 +32,8 @@
         android:layout_height="wrap_content"
         android:hint="@string/AnswerHere"
         android:layout_margin="10dp"
-        android:layout_below="@id/imageViewTest"/>
+        android:layout_below="@id/imageViewTest"
+        android:singleLine="true"
+        android:imeOptions="actionNext"/>
 
 </RelativeLayout>
